Add option to use detecto API & model for predictions
This adds support, through a rather dumb mechanism, for using the
Detecto API (and the model I built for use with it) for prediction.

Because Detecto has a different API from plain ol' Torch, I've wrapped
the two methods of predictions in their own functions, and used a
boolean switch to pick which one I use.

This is by no means ideal, but for half an hour of hacking I'm further
ahead than I was.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>